+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Instant boot support using the `GameBoy.boot()` method
+* Support for pending cycles in web version
 
 ### Changed